move code from server-ct -> server
