+++ packages/server/lib/project-base.ts
@@ -33,6 +33,7 @@ import { SpecsStore } from './specs-store'
 import { createRoutes as createE2ERoutes } from './routes'
 import { checkSupportFile } from './project_utils'
 import type { LaunchArgs } from './open_project'
+import { createRoutes as createRoutesCT } from './routes-ct'
 
 // Cannot just use RuntimeConfigOptions as is because some types are not complete.
 // Instead, this is an interface of values that have been manually validated to exist
@@ -219,7 +220,7 @@ export class ProjectBase<TServer extends ServerE2E | ServerCt> extends EE {
       shouldCorrelatePreRequests: this.shouldCorrelatePreRequests,
       testingType: this.testingType,
       SocketCtor: this.testingType === 'e2e' ? SocketE2E : SocketCt,
-      createRoutes: createE2ERoutes, //  : createCTRoutes,
+      createRoutes: this.testingType === 'e2e' ? createE2ERoutes : createRoutesCT,
       specsStore,
     })
