+++ packages/server/lib/server-base.ts
@@ -26,7 +26,7 @@ import { allowDestroy, DestroyableHttpServer } from './util/server_destroy'
 import { SocketAllowed } from './util/socket_allowed'
 import { createInitialWorkers } from '@packages/rewriter'
 import type { SpecsStore } from './specs-store'
-import type { InitializeRoutes } from '../../server-ct/src/routes-ct'
+import type { InitializeRoutes } from './routes-ct'
 import type { Cfg } from './project-base'
 import type { Browser } from '@packages/server/lib/browsers/types'
 import type { ParsedHost } from '@packages/network/lib/cors'
